eliminate seh frame overhead in network guid unpacking and fix poly clip count type
Calling __try in Hooked_GetWowGUID forced msvc to generate seh registration frames and stack cookies on every incoming network guid decode, impacting packet processing during heavy combat log and entity spawn activity. Validating the self, outGuid, and buffer pointer ranges directly removes the need for exception handling while preserving protection against unmapped access outside the window bounds. Hook installation now verifies the target prologue bytes before detouring 0x0076DC20. Additionally, polygon clipping now writes zero vertex count as an integer rather than float.

+++ src/simd_math/collision_poly_clip_sse2.cpp
@@ -219,7 +219,7 @@ bool __cdecl Handle(const float* plane, float* poly, uint32_t value) {
     const Outcome mine = Decide(plane, poly, n);
     if (mine == kNoChange) { ++g_inside; return false; }
     if (mine == kEmpty) {
-        *(float*)((char*)poly + kCountOffset) = 0.0f;
+        *(int*)((char*)poly + kCountOffset) = 0;
         ++g_outside;
         return false;
     }
